+++ src/main/java/cl/duoc/demoreclamos/repository/ReclamosRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import cl.duoc.demoreclamos.model.Reclamos;
 
-public interface ReclamosRepository extends JpaRepository <Reclamos, Long>{
+public interface ReclamosRepository extends JpaRepository <Reclamos, Integer>{
 
 }
